requirements.txt
PyYAML
quamash
-e git://github.com/pyqtgraph/pyqtgraph.git#egg=pyqtgraph
picamera
